created MainFunctionsFile in config_parser
