+++ src/config_parser/ValidationConfigFile.cpp
@@ -106,41 +106,3 @@ bool firstNonAlNumChar(const std::string &line, int *lineCont, const std::string
     }
     return false;
 }
-bool validationConfigFile(const std::string &filePath)
-{
-    std::ifstream file(filePath.c_str());
-    if (!file.is_open())
-        throw std::runtime_error("❌ No se pudo abrir el archivo");
-    std::string line;
-    int contOpenKey = 0;
-    int contCloseKey = 0;
-    int firstOpenKey = 0;
-    int lastCloseKey = 0;
-    std::vector<std::string> lines;
-    int lineCont = 1;
-    while (std::getline(file, line))
-    {
-        if (isEmptyBraceOrSemicolonLine(line, &lineCont, filePath))
-            return false;
-        if (incorrectLineTermination(line, &lineCont, filePath))
-            return false;
-        if (firstNonAlNumChar(line, &lineCont, filePath))
-            return false;
-
-        processConfigLine(line, &lineCont, &contOpenKey, &contCloseKey, &firstOpenKey,
-                          &lastCloseKey);
-        ////ZONA DE CURRELE///
-        // repeticiones de name o value???
-        // TODO: MEJORAR EL TRIM PARA ESPACIOS POSTERIORES
-
-        //////////////////////
-        lines.push_back(line);
-        lineCont++;
-    }
-    if (resultProcesConfigLine(contOpenKey, contCloseKey, firstOpenKey, lastCloseKey, filePath))
-        return false;
-    // size_t i = -1;
-    // while (++i < lines.size())
-    //     std::cout << "Line[" << i << "]:" << lines[i] << std::endl;
-    return true;
-}
