+++ src/config_parser/test.conf
@@ -1,8 +1,7 @@
 server{
     listen 9090 8080;
-    server_name localhost;
+    server_name && localhost;
     location {
         root ./html;
         }
-        $
 }
